==> 2-Refine.Rmd <==
---
title: "Refine the Cars Data"
author: "Amit Kapoor"
date: "July 2015"
output: html_document
---

# Refine the Data

Now that we have got the raw data. We need to do a number of clean ups to actually make it into a useful data frame format for doing our exploration.

```{r}
library(dplyr)
library(stringr)

df.messy <- read.csv("data/cars.spec.raw.csv", stringsAsFactors = FALSE)
str(df.messy)
```

Lets make this from long to wide format - mainly for the specs

```{r}
library(tidyr)
df.tidier <- df.messy %>%
             separate(spec, c("spec", "value"), ":")

# Understand the number of specs - only 56
df.tidier %>% 
  group_by(spec) %>%
  summarize(n())

df.tidy <- df.tidier %>%
           spread(spec, value)
```

Lets fix the names in the columns

```{r}
tidy.names = c("name", "model", "url", "price", "type", "abs", 
               "acc", "aircon", "audiostr", "audio", "bluetooth", "bumper",
               "boot", "fbrake", "rbrake", "cdplayer", "locking",
               "airfilter", "cruise", "doorajar", "dseatwarn", 
               "airbags", "ebd", "disp", "engtype", "ftrack","ftyres",
               "fullsys", "tank", "fuel",
               "wheelcaps", "weight", "clearance", "heater",
               "immobiliser", "kerb", "keyless", "power", "torque",
               "kmplc", "kmplh", "cylinder", "size", 
               "sunroof", "pairbags", "steering", "fpowerwin", 
               "rpowerwin", "rainwiper", "rtrack", "rtyres", "seat", 
               "speaker", "tilt", "speed", "traction", "trans", 
               "tubeless", "turning", "usb", "wheelbase")

names(df.tidy) <- tidy.names
str(df.tidy)
```


Let's start with cleaning name to get brand names

```{r}
# Lets start with name 
df.tidy %>% 
  group_by(name) %>%
  summarize(n())

# Fix specific brands
df.tidy$name[df.tidy$name == "New Chevrolet Beat"] <- "Chevrolet Beat"
df.tidy$name[df.tidy$name == "New 4S Fluidic Hyundai Verna"] <- "Hyundai Verna"
df.tidy$name[df.tidy$name == "Chevrolet Sail Hatchback (U-VA)"] <- "Chevrolet Sail Hatchback"

# There are 203 unique names, lets start with taking the first word as brand
# In base R:  first <- unlist(strsplit(string, split=" "))[1]
df.tidy$brand <- word(df.tidy$name, 1)
df.tidy %>% 
  group_by(brand) %>%
  summarize(n()) %>%
  View()

str(df.tidy)
```

Now. lets get the price as a numeric feature - we need to remove "Price (INR)Rs. " and remove the commas and make it as numeric
```{r}
pr <- "Price (INR)Rs. "
len.pr <- str_length(pr)  # Can also use nchar()

df.tidy$price <- df.tidy$price %>% 
             str_sub(len.pr, nchar(df.tidy$price)) %>%
             str_replace_all(",", "") %>%
             as.numeric()
```

Now, lets get the body type - we need to remove "Body Type" and convert as factor

```{r}
bt <- "Body Type "
len.bt <- str_length(bt)  # Can also use nchar()

df.tidy$type <- df.tidy$type %>% 
             str_sub(len.bt, nchar(df.tidy$type)) 

df.tidy %>%
  group_by(type) %>%
  tally() %>%
  View()

# Clean up the tags

# Remove non ASCII character
df.tidy$type <- iconv(df.tidy$type, "latin1", "ASCII", sub="")

df.tidy$type[df.tidy$type == "Hatchabck"] <- "Hatchback"
df.tidy$type[df.tidy$type == "hatchback"] <- "Hatchback"
df.tidy$type[df.tidy$type == "Hatchbackâ€Ž"] <- "Hatchback"


df.tidy$type[df.tidy$name == "Audi Q7"] <- "MUV"

df.tidy$type[df.tidy$type == "Seadan"] <- "Sedan"
df.tidy$type[df.tidy$type == "Sedna"] <- "Sedan"
df.tidy$type[df.tidy$type == "Sedans"] <- "Sedan"
df.tidy$type[df.tidy$type == "Seadn"] <- "Sedan"
df.tidy$type[df.tidy$type == "Saloon"] <- "Sedan"
df.tidy$type[df.tidy$type == "Saloon Sedan"] <- "Sedan"
df.tidy$type[df.tidy$type == "Luxury Sedan"] <- "Sedan"
df.tidy$type[df.tidy$type == "Luxury"] <- "Sedan"


df.tidy$type[df.tidy$type == "VAN"] <- "MUV"
df.tidy$type[df.tidy$type == "Van"] <- "MUV"
df.tidy$type[df.tidy$type == "Cabriolet"] <- "Convertible"
df.tidy$type[df.tidy$type == "Sport Coupe"] <- "Convertible"
df.tidy$type[df.tidy$type == "Supercar"] <- "Convertible"
df.tidy$type[df.tidy$name == "Porsche Boxster"] <- "Convertible"

```

```{r}
# 
df.tidy <- df.tidy %>%
           separate(power, c("bhp", "rpm"), "@")

df.tidy$bhp <- df.tidy$bhp %>%
               str_replace_all("bhp", "") 
df.tidy$bhp <- as.numeric(df.tidy$bhp) 

df.tidy$bhp[df.tidy$bhp == 9066.00] <- 90.66

df.tidy$rpm <- df.tidy$rpm %>%
               str_replace_all("rpm", "") 

df.tidy$rpm <- df.tidy$rpm %>%
               str_replace_all("RPM", "") 

df.tidy$kmplh <- as.numeric(df.tidy$kmplh)


write.csv(df.tidy, file="data/cars_tidy.csv", quote = FALSE,
          row.names = FALSE)

```

```{r}
# Take Petrol + Hatchback & Sedan and Price < 9L

df.small <- df.tidy %>%
            filter(fuel == ' Petrol') %>%
            arrange(name, price) %>%
            filter(price < 1000000)  %>%
            mutate(price = round(price/1000, 0)) %>%
            filter(type =='Hatchback' | type == "Sedan") %>%
            select(name, price, type, kmplh, bhp, fuel, brand)

# Delete duplicate rows      
df.small <- df.small[!duplicated(df.small$name),]
# Delete NAs
df.small <- df.small[complete.cases(df.small),]
       
# Delete some duplicate models
df.small <- df.small[df.small$name != "Volkswagen Polo GT", ]
df.small <- df.small[df.small$name != "Volkswagen Cross Polo", ]
df.small <- df.small[df.small$name != "Maruti Suzuki WagonR Stingray", ]
df.small <- df.small[df.small$name != "Honda Jazz", ]
df.small <- df.small[df.small$name != "Toyota Etios Xclusive", ]
df.small <- df.small[df.small$name != "Hyundai Grand i10", ]
df.small <- df.small[df.small$name != "Hyundai i20 Active", ]


View(df.small)
write.csv(df.small, file="data/cars_small.csv", quote = FALSE,
          row.names = FALSE)
```
