==> 3-Explore.Rmd <==
---
title: "Explore the Cars"
author: "Amit Kapoor"
date: "July 2015"
output: html_document
---

Lets explore the data set now

```{r}
# Import the data
df.hatch <- df.small %>%
            filter(type == "Hatchback")

df.sedan <- df.small %>%
            filter(type == "Sedan")

# Explore
library(ggplot2)

## Price vs bhp
ggplot(df.small) + aes(bhp,price) + geom_point() +
  geom_smooth(method = "lm")

ggplot(df.small) + aes(bhp,price,color = type) + geom_point() +
  geom_smooth(method = "lm")

ggplot(df.hatch) + aes(bhp,price) + geom_point() + 
  geom_smooth(method = "lm")

ggplot(df.sedan) + aes(bhp,price) + geom_point() + 
  geom_smooth(method = "lm")

## Price vs Kmplh
ggplot(df.small) + aes(kmplh,price) + geom_point() +
  geom_smooth(method = "lm")

ggplot(df.small) + aes(kmplh,price,color = type) + geom_point() +
  geom_smooth(method = "lm")

ggplot(df.hatch) + aes(kmplh,price) + geom_point() + 
  geom_smooth(method = "lm")

ggplot(df.sedan) + aes(kmplh,price) + geom_point() + 
  geom_smooth(method = "lm")

```
